+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 All Notable changes to *jkphl/rdfa-lite-microdata* will be documented in this file. The format is based on [Keep a Changelog](http://keepachangelog.com/). This project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## [0.3.0] - unreleased
+## [0.3.0] - 2017-03-17
 ### Added
-* Support for enhanced property lists and component based property names
+* Support for enhanced property lists and property names
+* IRI mode (IRI based item types and property names)
 
 ## [0.2.1] - 2017-03-11
 ### Fixed